feat: implement resource mapping for all supporter monitored resource types (#460)
This may be considered a breaking change for trace exporter, as this
changes the logic to select a monitored resource for populating the
special `g.co/{type}/{label}` span labels. However, I don't expect this
to have a big impact on users and we never declared the resource
mapping/detection stable anyway.